# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionResult": {
        "properties": {
          "error_code": {
            "default": "",
            "maxLength": 100,
            "title": "Error Code",
            "type": "string"
          },
          "provider_guid": {
            "default": "",
            "maxLength": 255,
            "title": "Provider Guid",
            "type": "string"
          },
          "readback": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readback"
          },
          "status": {
            "pattern": "^(PROVIDER_ACCEPTED|FAILED|AMBIGUOUS)$",
            "title": "Status",
            "type": "string"
          },
          "voucher_number": {
            "default": "",
            "maxLength": 100,
            "title": "Voucher Number",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ActionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeRequest": {
        "properties": {
          "approval_data": {
            "additionalProperties": true,
            "title": "Approval Data",
            "type": "object"
          },
          "thread_id": {
            "maxLength": 255,
            "minLength": 8,
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "approval_data"
        ],
        "title": "ResumeRequest",
        "type": "object"
      },
...
      "TallyBridgeClaimRequest": {
        "properties": {
          "bridge_device_id": {
            "minLength": 1,
            "title": "Bridge Device Id",
            "type": "string"
          },
          "tally_company": {
            "minLength": 1,
            "title": "Tally Company",
            "type": "string"
          }
        },
        "required": [
          "tally_company",
          "bridge_device_id"
        ],
        "title": "TallyBridgeClaimRequest",
        "type": "object"
      },
      "TallyBridgeResultRequest": {
        "properties": {
          "bridge_device_id": {
            "title": "Bridge Device Id",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "status": {
            "enum": [
              "SUCCEEDED",
              "FAILED",
              "NEEDS_REVIEW"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "correlation_id",
          "bridge_device_id",
          "status"
        ],
        "title": "TallyBridgeResultRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "firmOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bridge/actions/claim": {
      "post": {
        "description": "Python bridge reads remain compatible; writes require the secure desktop agent.",
        "operationId": "claim_action_api_bridge_actions_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallyBridgeClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Action",
        "tags": [
          "tally-bridge-actions"
        ]
      }
    },
    "/api/bridge/actions/{action_id}/result": {
      "post": {
        "description": "Close an old lease safely, but never certify an unverified legacy write.",
        "operationId": "report_action_result_api_bridge_actions__action_id__result_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TallyBridgeResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Action Result",
        "tags": [
          "tally-bridge-actions"
        ]
      }
    },
...
    "/api/tally-agent/actions/claim": {
      "post": {
        "operationId": "claim_action_api_tally_agent_actions_claim_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FirmOS-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 36,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Device"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Firm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Firm"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Body-SHA256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Body-Sha256"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Action",
        "tags": [
          "tally-agent"
        ]
      }
    },
...
    "/api/tally-agent/actions/{action_id}/result": {
      "post": {
        "operationId": "report_result_api_tally_agent_actions__action_id__result_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 36,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Device"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Firm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Firm"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Body-SHA256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Body-Sha256"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Result",
        "tags": [
          "tally-agent"
        ]
      }
    },
...
    "/api/workflows/{workflow_id}/resume": {
      "post": {
        "description": "Resume only the caller's persistent workflow run.",
        "operationId": "resume_workflow_api_workflows__workflow_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}/run": {
      "post": {
        "description": "Start an owned workflow; user-supplied thread IDs never reach storage.",
        "operationId": "run_workflow_api_workflows__workflow_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thread_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 8,
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Initial State"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}/run/{thread_id}": {
      "get": {
        "description": "Read only the caller's persistent workflow run.",
        "operationId": "get_workflow_state_api_workflows__workflow_id__run__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow State",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}